Refactor - PR suggestions
